make navigation more robust, saveState = true was the main issue
The ai led me on the wron path, kept telling me i needed to remove navcontrollers from the view models (it can apparently cause memory leak on recomposition)
The root of the bug was that the devices page was saving the state of being in 'device settings'. So when I returned it still had the old state settings shown
mad ethe back handlers better, and removed all dependencies on the navcontroller from viewmodels, so all was not lost

main bug i was trying to solve
reproduction:
 open device page
 click settings cog (device settings appears)
 click hamburger menu and open any other page
 click hamburger menu and open devices page, its stuck on settings for some reason

It solved a bunch of other dirty hacks i had made for weird navigation issues, so as i said, all is not lost and it pronbably improved the code a bit. Although events make it harder to read now :(